Apply new layouts to `ManageAccount` and show key pages
