+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/evmprivatekey/EvmPrivateKeyModule.kt
@@ -1,32 +1,21 @@
-package io.horizontalsystems.bankwallet.modules.showkey
+package io.horizontalsystems.bankwallet.modules.evmprivatekey
 
-import android.os.Parcelable
 import androidx.core.os.bundleOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import io.horizontalsystems.bankwallet.core.App
 import io.horizontalsystems.bankwallet.entities.Account
-import kotlinx.parcelize.Parcelize
 
-object ShowKeyModule {
+object EvmPrivateKeyModule {
     const val ACCOUNT = "account"
 
     class Factory(private val account: Account) : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            return ShowKeyViewModel(account, App.pinComponent, App.evmBlockchainManager) as T
+            return EvmPrivateKeyViewModel(account, App.evmBlockchainManager) as T
         }
     }
 
     fun prepareParams(account: Account) = bundleOf(ACCOUNT to account)
-
-    @Parcelize
-    data class PrivateKey(val blockchain: String, val value: String) : Parcelable
-
-    enum class ViewState{
-        Warning, Key
-    }
-
-    data class WordNumbered(val word: String, val number: Int)
 
 }
