+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/publickeys/PublicKeysViewModel.kt
@@ -1,71 +1,26 @@
-package io.horizontalsystems.bankwallet.modules.showkey
+package io.horizontalsystems.bankwallet.modules.publickeys
 
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
-import io.horizontalsystems.bankwallet.core.managers.EvmBlockchainManager
 import io.horizontalsystems.bankwallet.entities.Account
 import io.horizontalsystems.bankwallet.entities.AccountType
 import io.horizontalsystems.bankwallet.entities.BitcoinCashCoinType
 import io.horizontalsystems.bitcoincash.MainNetBitcoinCash
 import io.horizontalsystems.bitcoinkit.MainNet
-import io.horizontalsystems.core.IPinComponent
 import io.horizontalsystems.dashkit.MainNetDash
-import io.horizontalsystems.ethereumkit.core.signer.Signer
-import io.horizontalsystems.ethereumkit.core.toHexString
 import io.horizontalsystems.hdwalletkit.HDKeychain
 import io.horizontalsystems.litecoinkit.MainNetLitecoin
-import io.horizontalsystems.marketkit.models.BlockchainType
 
-class ShowKeyViewModel(
+class PublicKeysViewModel(
     account: Account,
-    pinComponent: IPinComponent,
-    private val evmBlockchainManager: EvmBlockchainManager
 ) : ViewModel() {
-    private val words: List<String>
     private val seed: ByteArray?
 
-    var passphrase by mutableStateOf("")
-        private set
-
-    var viewState by mutableStateOf(ShowKeyModule.ViewState.Warning)
-        private set
-
-    var wordsNumbered by mutableStateOf<List<ShowKeyModule.WordNumbered>>(listOf())
-        private set
-
     init {
-        if (account.type is AccountType.Mnemonic) {
-            words = account.type.words
-            wordsNumbered = words.mapIndexed { index, word ->
-                ShowKeyModule.WordNumbered(word, index + 1)
-            }
-            passphrase = account.type.passphrase
-            seed = account.type.seed
+        seed = if (account.type is AccountType.Mnemonic) {
+            account.type.seed
         } else {
-            words = listOf()
-            seed = null
+            null
         }
-    }
-
-    val isPinSet = pinComponent.isPinSet
-
-    val ethereumPrivateKey: String
-        get() = Signer.privateKey(
-            words,
-            passphrase,
-            evmBlockchainManager.getChain(BlockchainType.Ethereum)
-        ).toByteArray().let {
-            if (it.size > 32) {
-                it.copyOfRange(1, it.size)
-            } else {
-                it
-            }.toHexString()
-        }
-
-    fun showKey() {
-        viewState = ShowKeyModule.ViewState.Key
     }
 
     fun bitcoinPublicKeys(derivation: AccountType.Derivation): String? {
@@ -80,7 +35,7 @@ class ShowKeyViewModel(
         seed ?: return null
         val keychain = HDKeychain(seed)
 
-        val network = when(coinType){
+        val network = when (coinType) {
             BitcoinCashCoinType.type0 -> MainNetBitcoinCash(MainNetBitcoinCash.CoinType.Type0)
             BitcoinCashCoinType.type145 -> MainNetBitcoinCash(MainNetBitcoinCash.CoinType.Type145)
         }
