+++ app/src/main/res/drawable/icon_link_20.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="20dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M16.606,3.394C14.747,1.535 11.733,1.535 9.875,3.394L8.255,5.014C7.962,5.307 7.962,5.782 8.255,6.075C8.547,6.368 9.022,6.368 9.315,6.075L10.935,4.455C12.208,3.182 14.272,3.182 15.545,4.455C16.818,5.728 16.818,7.792 15.545,9.065L13.925,10.685C13.632,10.978 13.632,11.453 13.925,11.745C14.218,12.038 14.693,12.038 14.986,11.745L16.606,10.125C18.465,8.267 18.465,5.253 16.606,3.394ZM6.075,9.315C6.368,9.022 6.368,8.547 6.075,8.255C5.782,7.962 5.307,7.962 5.014,8.255L3.394,9.875C1.535,11.733 1.535,14.747 3.394,16.606C5.253,18.465 8.267,18.465 10.125,16.606L11.745,14.986C12.038,14.693 12.038,14.218 11.745,13.925C11.453,13.632 10.978,13.632 10.685,13.925L9.065,15.545C7.792,16.818 5.728,16.818 4.455,15.545C3.182,14.272 3.182,12.208 4.455,10.935L6.075,9.315ZM7.22,12.78C7.513,13.073 7.988,13.073 8.28,12.78L12.78,8.28C13.073,7.987 13.073,7.513 12.78,7.22C12.487,6.927 12.013,6.927 11.72,7.22L7.22,11.72C6.927,12.013 6.927,12.488 7.22,12.78Z"
+      android:fillColor="#808085"
+      android:fillType="evenOdd"/>
+</vector>
