+++ app/src/main/res/drawable/icon_warning_2_20.xml
@@ -1,0 +1,17 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="20dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M9.997,5.5C10.411,5.5 10.747,5.836 10.747,6.25V9.988C10.747,10.403 10.411,10.738 9.997,10.738C9.582,10.738 9.247,10.403 9.247,9.988V6.25C9.247,5.836 9.582,5.5 9.997,5.5Z"
+      android:fillColor="#808085"
+      android:fillType="evenOdd"/>
+  <path
+      android:pathData="M8.997,12.988C8.997,13.541 9.444,13.988 9.997,13.988C10.549,13.988 10.997,13.541 10.997,12.988C10.997,12.436 10.549,11.988 9.997,11.988C9.444,11.988 8.997,12.436 8.997,12.988Z"
+      android:fillColor="#808085"/>
+  <path
+      android:pathData="M17.281,13.226L11.287,3.23C10.704,2.258 9.296,2.258 8.713,3.23L2.719,13.226C2.119,14.226 2.839,15.499 4.006,15.499H15.994C17.161,15.499 17.881,14.226 17.281,13.226ZM12.574,2.458C11.408,0.514 8.592,0.514 7.426,2.458L1.432,12.453C0.232,14.454 1.673,17 4.006,17H15.994C18.327,17 19.768,14.454 18.568,12.453L12.574,2.458Z"
+      android:fillColor="#808085"
+      android:fillType="evenOdd"/>
+</vector>
